requirements.txt
# Core
pandas
numpy
scikit-learn
xgboost
lightgbm
pyarrow
joblib

# NBA Data & API
nba_api
requests
python-dotenv

# Google Cloud Platform
google-cloud-storage
google-cloud-bigquery
google-api-core
google-cloud-core

# ML & Experimentation
mlflow
optuna

# Server & Deployment
flask
gunicorn
